remove notebook diff and refactor quant cfg util
